fix typo

+++ entityx/Entity.h
@@ -35,7 +35,7 @@
 namespace entityx {
 
 typedef std::uint32_t uint32_t;
-typedef std::uint64_t uint32_t;
+typedef std::uint64_t uint64_t;
 
 class EntityManager;
